use plain js for echo server background page
